CRF-Feat: service to create transaction

+++ src/main/java/com/perfums/transactions/domain/dto/FragranceSizeDTO.java
@@ -3,6 +3,8 @@ package com.perfums.transactions.domain.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class FragranceSizeDTO {
@@ -10,7 +12,7 @@ public class FragranceSizeDTO {
     private String unit;
     private String label;
     private Double size;
-    private Double originalPrice;
+    private BigDecimal originalPrice;
     private Double price;
     private Integer stock;
     private String imageUrl;
